Fix generator-produced warnings: dead keepArea var and drop-control casts
parseboxstructure.py's saveoneColumnMultiRows declared an unused
keepArea (leftover from the '+' layout's pattern, where it's actually
used) - removed for every |=N layout, regenerating
maxdiffuser/sampleplayer/minireverb's Editor.h.

The 'drop' control lambda always cast to int, but some hand-written
setters take size_t with no clamp. Added an optional "signed" blueprint
field (default true, matching today's int cast) so the generator can
match each setter's actual type; set it false on sampleplayer's
type/solo, metronome's dropBars, minireverb's order, and
sampleplayertimestretched's type. delay::setSyncDivision and
metronome::setPreset keep their existing clamped int path untouched.

+++ examples/metronome/src/MetronomeProcessor.h
@@ -351,7 +351,7 @@ class AudioPluginAudioProcessor : public juce::AudioProcessor, public juce::Audi
             {"dropBars",
              [](AudioPluginAudioProcessor& p, const float v)
              {
-                 p.pluginRunner->setDropBars(static_cast<int>(v));
+                 p.pluginRunner->setDropBars(static_cast<size_t>(v));
                  p.m_fileIo.updateParameter(PatchParameters::Id::dropBars, v);
              }},
             {"metroVolume",
